* Moved some files to libtomahawk.
