+++ src/libtomahawk/accounts/AccountFactoryWrapperDelegate.cpp
