+++ src/libtomahawk/accounts/AccountFactoryWrapperDelegate.h
