+++ src/libtomahawk/accounts/DelegateConfigWrapper.cpp
