+++ src/libtomahawk/accounts/DelegateConfigWrapper.h
